Map marker label enhancer no color props (#4772)
* chore(map-marker): remove color properties

* chore(map-marker): extract needle and drag shadow components
